test: split readme-examples into per-package files
Move core-only README example tests to packages/core/src/readme-examples.test.ts.
Keep root-README signup-form test in packages/validators/src/readme-examples.test.ts.
ignoreValidationResult test rewritten to use a custom validator so the new
core test file imports only @form-validator-js/core.

Two tests (validation contexts, per-field error message overrides) are dropped
here and re-added in the next commit with custom-validator examples that match
the rewritten core README.
